requirements.txt
absl-py==0.7.0
ansiwrap==0.8.4
astor==0.7.1
attrs==18.2.0
backcall==0.1.0
biopython==1.73
bleach==1.5.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
future==0.17.1
gast==0.2.2
grpcio==1.19.0
html5lib==0.9999999
idna==2.8
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.2
opencv-python==4.0.0.21
pandas==0.24.1
pandocfilters==1.4.2
papermill==0.19.0
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.6.1
psycopg2==2.7.7
ptyprocess==0.6.0
py==1.8.0
Pygments==2.3.1
pymssql==2.1.4
PyMySQL==0.9.3
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
python-hostlist==1.18
python-swiftclient==3.6.0
pytz==2018.9
PyYAML==3.13
pyzmq==18.0.0
qtconsole==4.4.3
requests==2.21.0
retry==0.9.2
sci==0.1.3
scikit-learn==0.20.2
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
tensorflow-probability==0.6.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
textwrap3==0.9.2
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.4.2
